Gather resmap factory code into one file.
